backend/requirements.txt
Django==5.1.6
pymongo==4.11.1
djangorestframework
mongoengine